* Double the retries in CI
